Seahawks-Monitoring/seahawks_harvester/requirements.txt
flask
python-nmap
requests
